[tests] Clean up test runners
