--- tests/sdk/nodejs/autonaming/autonaming_test.go
@@ -1,155 +1,0 @@
-// Copyright 2016-2019, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ints
-
-import (
-	"strings"
-	"testing"
-
-	"github.com/pulumi/pulumi-kubernetes/provider/v3/pkg/openapi"
-	"github.com/pulumi/pulumi-kubernetes/tests/v3"
-	"github.com/pulumi/pulumi/pkg/v3/resource/deploy/providers"
-	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
-	"github.com/stretchr/testify/assert"
-)
-
-var step1Name interface{}
-var step2Name interface{}
-var step3Name interface{}
-
-func TestAutonaming(t *testing.T) {
-	integration.ProgramTest(t, &integration.ProgramTestOptions{
-		Dir:          "step1",
-		Dependencies: []string{"@pulumi/kubernetes"},
-		Quick:        true,
-		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-			assert.NotNil(t, stackInfo.Deployment)
-			assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-			tests.SortResourcesByURN(stackInfo)
-
-			stackRes := stackInfo.Deployment.Resources[3]
-			assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-			provRes := stackInfo.Deployment.Resources[2]
-			assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-			//
-			// Assert Pod is successfully given a unique name by Pulumi.
-			//
-
-			pod := stackInfo.Deployment.Resources[1]
-			assert.Equal(t, "autonaming-test", string(pod.URN.Name()))
-			step1Name, _ = openapi.Pluck(pod.Outputs, "metadata", "name")
-			assert.True(t, strings.HasPrefix(step1Name.(string), "autonaming-test-"))
-
-			autonamed, _ := openapi.Pluck(pod.Outputs, "metadata", "annotations", "pulumi.com/autonamed")
-			assert.Equal(t, "true", autonamed)
-		},
-		EditDirs: []integration.EditDir{
-			{
-				Dir:      "step2",
-				Additive: true,
-				ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-					assert.NotNil(t, stackInfo.Deployment)
-					assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-					tests.SortResourcesByURN(stackInfo)
-
-					stackRes := stackInfo.Deployment.Resources[3]
-					assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-					provRes := stackInfo.Deployment.Resources[2]
-					assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-					//
-					// Assert Pod was replaced, i.e., destroyed and re-created, with allocating a new name.
-					//
-
-					pod := stackInfo.Deployment.Resources[1]
-					assert.Equal(t, "autonaming-test", string(pod.URN.Name()))
-					step2Name, _ = openapi.Pluck(pod.Outputs, "metadata", "name")
-					assert.True(t, strings.HasPrefix(step2Name.(string), "autonaming-test-"))
-
-					autonamed, _ := openapi.Pluck(pod.Outputs, "metadata", "annotations", "pulumi.com/autonamed")
-					assert.Equal(t, "true", autonamed)
-
-					assert.NotEqual(t, step1Name, step2Name)
-
-				},
-			},
-			{
-				Dir:      "step3",
-				Additive: true,
-				ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-					assert.NotNil(t, stackInfo.Deployment)
-					assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-					tests.SortResourcesByURN(stackInfo)
-
-					stackRes := stackInfo.Deployment.Resources[3]
-					assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-					provRes := stackInfo.Deployment.Resources[2]
-					assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-					//
-					// Assert Pod was NOT replaced, and has the same name, previously allocated by Pulumi.
-					//
-
-					pod := stackInfo.Deployment.Resources[1]
-					assert.Equal(t, "autonaming-test", string(pod.URN.Name()))
-					step3Name, _ = openapi.Pluck(pod.Outputs, "metadata", "name")
-					assert.True(t, strings.HasPrefix(step3Name.(string), "autonaming-test-"))
-
-					autonamed, _ := openapi.Pluck(pod.Outputs, "metadata", "annotations", "pulumi.com/autonamed")
-					assert.Equal(t, "true", autonamed)
-
-					assert.Equal(t, step2Name, step3Name)
-				},
-			},
-			{
-				Dir:      "step4",
-				Additive: true,
-				ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-					assert.NotNil(t, stackInfo.Deployment)
-					assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-					tests.SortResourcesByURN(stackInfo)
-
-					stackRes := stackInfo.Deployment.Resources[3]
-					assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-					provRes := stackInfo.Deployment.Resources[2]
-					assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-					//
-					// User has specified their own name for the Pod, so we replace it, and Pulumi does NOT
-					// allocate a name on its own.
-					//
-
-					pod := stackInfo.Deployment.Resources[1]
-					assert.Equal(t, "autonaming-test", string(pod.URN.Name()))
-					name, _ := openapi.Pluck(pod.Outputs, "metadata", "name")
-					assert.Equal(t, "autonaming-test", name.(string))
-
-					_, autonamed := openapi.Pluck(pod.Outputs, "metadata", "annotations", "pulumi.com/autonamed")
-					assert.False(t, autonamed)
-				},
-			},
-		},
-	})
-}
