--- tests/sdk/nodejs/delete-before-replace/delete_before_replace_test.go
@@ -1,133 +1,0 @@
-// Copyright 2016-2019, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ints
-
-import (
-	"testing"
-
-	"github.com/pulumi/pulumi-kubernetes/provider/v3/pkg/openapi"
-	"github.com/pulumi/pulumi-kubernetes/tests/v3"
-	"github.com/pulumi/pulumi/pkg/v3/resource/deploy/providers"
-	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestPod(t *testing.T) {
-	integration.ProgramTest(t, &integration.ProgramTestOptions{
-		Dir:          "step1",
-		Dependencies: []string{"@pulumi/kubernetes"},
-		Quick:        true,
-		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-			assert.NotNil(t, stackInfo.Deployment)
-			assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-			tests.SortResourcesByURN(stackInfo)
-
-			stackRes := stackInfo.Deployment.Resources[3]
-			assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-			provRes := stackInfo.Deployment.Resources[2]
-			assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-			//
-			// Assert pod is successfully created.
-			//
-
-			pod := stackInfo.Deployment.Resources[1]
-			name, _ := openapi.Pluck(pod.Outputs, "metadata", "name")
-			assert.Equal(t, name.(string), "pod-test")
-
-			// Not autonamed.
-			_, autonamed := openapi.Pluck(pod.Outputs, "metadata", "annotations", "pulumi.com/autonamed")
-			assert.False(t, autonamed)
-
-			// Status is "Running"
-			phase, _ := openapi.Pluck(pod.Outputs, "status", "phase")
-			assert.Equal(t, "Running", phase)
-
-			// Status "Ready" is "True".
-			conditions, _ := openapi.Pluck(pod.Outputs, "status", "conditions")
-			ready := conditions.([]interface{})[1].(map[string]interface{})
-			readyType := ready["type"]
-			assert.Equal(t, "Ready", readyType)
-			readyStatus := ready["status"]
-			assert.Equal(t, "True", readyStatus)
-
-			// Container is called "nginx" and uses image "nginx:1.13-alpine".
-			containerStatuses, _ := openapi.Pluck(pod.Outputs, "status", "containerStatuses")
-			containerStatus := containerStatuses.([]interface{})[0].(map[string]interface{})
-			containerName := containerStatus["name"]
-			assert.Equal(t, "nginx", containerName)
-			image := containerStatus["image"]
-			assert.Equal(t, "nginx:1.13-alpine", image)
-		},
-		EditDirs: []integration.EditDir{
-			{
-				Dir:      "step2",
-				Additive: true,
-				ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-					assert.NotNil(t, stackInfo.Deployment)
-					assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-					tests.SortResourcesByURN(stackInfo)
-
-					stackRes := stackInfo.Deployment.Resources[3]
-					assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-					provRes := stackInfo.Deployment.Resources[2]
-					assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-					//
-					// Assert Pod is deleted before being replaced with the new Pod, running
-					// nginx:1.15-alpine.
-					//
-					// Because the Pod name is supplied in the resource definition, we are forced to delete it
-					// before replacing it, as otherwise Kubernetes would complain that it can't have two Pods
-					// with the same name.
-					//
-
-					pod := stackInfo.Deployment.Resources[1]
-					name, _ := openapi.Pluck(pod.Outputs, "metadata", "name")
-					assert.Equal(t, name.(string), "pod-test")
-
-					// Not autonamed.
-					_, autonamed := openapi.Pluck(pod.Outputs, "metadata", "annotations", "pulumi.com/autonamed")
-					assert.False(t, autonamed)
-
-					// Status is "Running"
-					phase, _ := openapi.Pluck(pod.Outputs, "status", "phase")
-					assert.Equal(t, "Running", phase)
-
-					// Status "Ready" is "True".
-					conditions, _ := openapi.Pluck(pod.Outputs, "status", "conditions")
-					ready := conditions.([]interface{})[1].(map[string]interface{})
-					readyType := ready["type"]
-					assert.Equal(t, "Ready", readyType)
-					readyStatus := ready["status"]
-					assert.Equal(t, "True", readyStatus)
-
-					// Container is called "nginx" and uses image "nginx:1.13-alpine".
-					containerStatuses, _ := openapi.Pluck(pod.Outputs, "status", "containerStatuses")
-					containerStatus := containerStatuses.([]interface{})[0].(map[string]interface{})
-					containerName := containerStatus["name"]
-					assert.Equal(t, "nginx", containerName)
-					image := containerStatus["image"]
-					assert.Equal(t, "nginx:1.15-alpine", image)
-				},
-			},
-		},
-	})
-}
