--- tests/sdk/nodejs/empty-array/step2/tsconfig.json
@@ -1,22 +1,0 @@
-{
-    "compilerOptions": {
-        "outDir": "bin",
-        "target": "es6",
-        "module": "commonjs",
-        "moduleResolution": "node",
-        "declaration": true,
-        "sourceMap": true,
-        "stripInternal": true,
-        "experimentalDecorators": true,
-        "pretty": true,
-        "noFallthroughCasesInSwitch": true,
-        "noImplicitAny": true,
-        "noImplicitReturns": true,
-        "forceConsistentCasingInFileNames": true,
-        "strictNullChecks": true
-    },
-    "files": [
-        "index.ts"
-    ]
-}
-
