--- tests/sdk/nodejs/namespace/step2/Pulumi.yaml
@@ -1,3 +1,0 @@
-name: namespace-tests
-description: Tests whether Pulumi successfully updates namespace resources.
-runtime: nodejs
