--- tests/sdk/nodejs/namespace/step2/package.json
@@ -1,13 +1,0 @@
-{
-    "name": "steps",
-    "version": "0.1.0",
-    "dependencies": {
-        "@pulumi/pulumi": "latest",
-        "@pulumi/random": "^2.1.1"
-    },
-    "devDependencies": {
-    },
-    "peerDependencies": {
-        "@pulumi/kubernetes": "latest"
-    }
-}
