--- tests/sdk/nodejs/query/query_test.go
@@ -1,81 +1,0 @@
-// Copyright 2016-2019, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ints
-
-import (
-	"testing"
-
-	"github.com/pulumi/pulumi-kubernetes/tests/v3"
-	"github.com/pulumi/pulumi/pkg/v3/resource/deploy/providers"
-	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/resource"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestQuery(t *testing.T) {
-	integration.ProgramTest(t, &integration.ProgramTestOptions{
-		Dir:          "step1",
-		Dependencies: []string{"@pulumi/kubernetes"},
-		Quick:        true,
-		StackName:    "query-test-c186bcc3-1572-44d8-b7d5-1028853682c3", // Chosen by fair dice roll. Guaranteed to be random.
-		CloudURL:     "file://~",                                        // Required; we hard-code the stack name
-		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-			assert.NotNil(t, stackInfo.Deployment)
-			assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-			tests.SortResourcesByURN(stackInfo)
-
-			stackRes := stackInfo.Deployment.Resources[3]
-			assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-			provRes := stackInfo.Deployment.Resources[2]
-			assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-			//
-			// Assert Pod is successfully given a unique name by Pulumi.
-			//
-
-			pod := stackInfo.Deployment.Resources[1]
-			assert.Equal(t, "query-test", string(pod.URN.Name()))
-		},
-		EditDirs: []integration.EditDir{
-			{
-				Dir:       "step2",
-				Additive:  true,
-				QueryMode: true,
-				ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-					//
-					// Verify no resources were deleted.
-					//
-					assert.NotNil(t, stackInfo.Deployment)
-					assert.Equal(t, 4, len(stackInfo.Deployment.Resources))
-
-					tests.SortResourcesByURN(stackInfo)
-
-					stackRes := stackInfo.Deployment.Resources[3]
-					assert.Equal(t, resource.RootStackType, stackRes.URN.Type())
-
-					provRes := stackInfo.Deployment.Resources[2]
-					assert.True(t, providers.IsProviderType(provRes.URN.Type()))
-
-					//
-					// If we pass this point, the query did NOT throw an error, and is therefore
-					// successful.
-					//
-				},
-			},
-		},
-	})
-}
