--- tests/sdk/nodejs/secrets/secrets_test.go
@@ -1,47 +1,0 @@
-// Copyright 2016-2019, Pulumi Corporation.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package ints
-
-import (
-	b64 "encoding/base64"
-	json "encoding/json"
-	"testing"
-
-	"github.com/pulumi/pulumi/pkg/v3/testing/integration"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestSecrets(t *testing.T) {
-	secretMessage := "secret message for testing"
-
-	integration.ProgramTest(t, &integration.ProgramTestOptions{
-		Dir:          "step1",
-		Dependencies: []string{"@pulumi/kubernetes"},
-		Config: map[string]string{
-			"message": secretMessage,
-		},
-		ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
-			assert.NotNil(t, stackInfo.Deployment)
-			state, err := json.Marshal(stackInfo.Deployment)
-			assert.NoError(t, err)
-
-			assert.NotContains(t, string(state), secretMessage)
-
-			// The program converts the secret message to base64, to make a ConfigMap from it, so the state
-			// should also not contain the base64 encoding of secret message.
-			assert.NotContains(t, string(state), b64.StdEncoding.EncodeToString([]byte(secretMessage)))
-		},
-	})
-}
